internal/config: new dynamic package for configdynamic func call
